+++ timetable_solutions/domain/data_upload_processing/processors/_other_models.py
@@ -3,7 +3,7 @@ Subclasses of the BaseFileUploadProcessor used to handle the upload of individua
 """
 
 # Local application imports
-from domain.data_upload_processing.processors.base_processors import (
+from domain.data_upload_processing.processors._base import (
     BaseFileUploadProcessor,
 )
 from domain.data_upload_processing.constants import UploadFileStructure
